requirements.txt
async-google-trans-new==1.4.6
gTTS==2.3.2
playsound==1.3.0
deepl-translate==1.2.0
twitchio==2.7.0
emoji==2.8.0
python-dotenv==1.0.0
pywin32;platform_system == 'Windows'
uwuify
r20

